Bumped versions of dependencies to the latest versions which are also available for download on bintray. I should clean my local maven repo more often to find if I'm using non downloadable dependencies :-).
Modified the *pom\_add\_snapshot\_to\_version.bsh* and the *pom-change\_version.bsh* scripts to assume that each sub module have the same version as the parent and that for submodules only the parent version is specified. This is the most common case as I see it. Thereby the _modifyParent_ parameter is no longer valid (just ignored).